Clean: rename and split funcs from cmd_path_and_ecevce_cmd.c file
